meteo: refactor: Update heading and ground_track functions to return u32 instead of f64

+++ src/adsb/heading.rs
@@ -1,6 +1,6 @@
-pub fn heading(message: &[u32]) -> Option<f64> {
+pub fn heading(message: &[u32]) -> Option<u32> {
     if let Some((_, value)) = crate::adsb::flag_and_range_value(message, 1, 47, 56) {
-        Some(((value * 360) >> 10) as f64)
+        Some((value * 360) >> 10)
     } else {
         None
     }
